Pv3Rs: Estimate the Cause of Recurrent Vivax Malaria using Genetic Data

Compute per-person posterior probabilities that each recurrent Plasmodium vivax (Pv) recurrence is a recrudescence, relapse, or reinfection (3Rs) using per-person P. vivax genetic data on two or more episodes.

See also

Author

Maintainer: Aimee Taylor aimee.taylor@pasteur.fr (ORCID)

Authors:

Yong See Foo

Other contributors:

Mans Magnusson [contributor]

l'INSTITUT PASTEUR [copyright holder]

European Union, Project 101110393 [funder]