List all allelic draws for three half siblings

Given a specified number of alleles for a single marker, generate_halfsib_alleles() enumerates all the ways three half siblings can draw alleles from their respective parents by firstly enumerating all allelic combinations for the parents, and by secondly enumerating all the inheritable combinations for the children.

Usage

enumerate_halfsib_alleles(n_alleles)

Arguments

n_alleles

Positive whole number specifying a per-marker number of alleles, otherwise known as marker cardinality.

Value

A character matrix. Each column is an individual. Each row is a possible allelic draw. Alleles are represented by the first n_alleles letters of the latin alphabet.

Examples

enumerate_halfsib_alleles(3)
#>        Half sibling 1 Half sibling 2 Half sibling 3
#>   [1,] "A"            "A"            "A"           
#>   [2,] "B"            "B"            "A"           
#>   [3,] "C"            "C"            "A"           
#>   [4,] "A"            "A"            "B"           
#>   [5,] "B"            "B"            "B"           
#>   [6,] "C"            "C"            "B"           
#>   [7,] "A"            "A"            "C"           
#>   [8,] "B"            "B"            "C"           
#>   [9,] "C"            "C"            "C"           
#>  [10,] "A"            "A"            "A"           
#>  [11,] "B"            "B"            "A"           
#>  [12,] "C"            "C"            "A"           
#>  [13,] "A"            "A"            "B"           
#>  [14,] "B"            "B"            "B"           
#>  [15,] "C"            "C"            "B"           
#>  [16,] "A"            "A"            "C"           
#>  [17,] "B"            "B"            "C"           
#>  [18,] "C"            "C"            "C"           
#>  [19,] "A"            "A"            "A"           
#>  [20,] "B"            "B"            "A"           
#>  [21,] "C"            "C"            "A"           
#>  [22,] "A"            "A"            "B"           
#>  [23,] "B"            "B"            "B"           
#>  [24,] "C"            "C"            "B"           
#>  [25,] "A"            "A"            "C"           
#>  [26,] "B"            "B"            "C"           
#>  [27,] "C"            "C"            "C"           
#>  [28,] "A"            "A"            "A"           
#>  [29,] "A"            "B"            "A"           
#>  [30,] "A"            "C"            "A"           
#>  [31,] "B"            "A"            "B"           
#>  [32,] "B"            "B"            "B"           
#>  [33,] "B"            "C"            "B"           
#>  [34,] "C"            "A"            "C"           
#>  [35,] "C"            "B"            "C"           
#>  [36,] "C"            "C"            "C"           
#>  [37,] "A"            "A"            "A"           
#>  [38,] "A"            "B"            "A"           
#>  [39,] "A"            "C"            "A"           
#>  [40,] "B"            "A"            "B"           
#>  [41,] "B"            "B"            "B"           
#>  [42,] "B"            "C"            "B"           
#>  [43,] "C"            "A"            "C"           
#>  [44,] "C"            "B"            "C"           
#>  [45,] "C"            "C"            "C"           
#>  [46,] "A"            "A"            "A"           
#>  [47,] "A"            "B"            "A"           
#>  [48,] "A"            "C"            "A"           
#>  [49,] "B"            "A"            "B"           
#>  [50,] "B"            "B"            "B"           
#>  [51,] "B"            "C"            "B"           
#>  [52,] "C"            "A"            "C"           
#>  [53,] "C"            "B"            "C"           
#>  [54,] "C"            "C"            "C"           
#>  [55,] "A"            "A"            "A"           
#>  [56,] "B"            "A"            "A"           
#>  [57,] "C"            "A"            "A"           
#>  [58,] "A"            "A"            "B"           
#>  [59,] "B"            "A"            "B"           
#>  [60,] "C"            "A"            "B"           
#>  [61,] "A"            "A"            "C"           
#>  [62,] "B"            "A"            "C"           
#>  [63,] "C"            "A"            "C"           
#>  [64,] "A"            "B"            "A"           
#>  [65,] "B"            "B"            "A"           
#>  [66,] "C"            "B"            "A"           
#>  [67,] "A"            "B"            "B"           
#>  [68,] "B"            "B"            "B"           
#>  [69,] "C"            "B"            "B"           
#>  [70,] "A"            "B"            "C"           
#>  [71,] "B"            "B"            "C"           
#>  [72,] "C"            "B"            "C"           
#>  [73,] "A"            "C"            "A"           
#>  [74,] "B"            "C"            "A"           
#>  [75,] "C"            "C"            "A"           
#>  [76,] "A"            "C"            "B"           
#>  [77,] "B"            "C"            "B"           
#>  [78,] "C"            "C"            "B"           
#>  [79,] "A"            "C"            "C"           
#>  [80,] "B"            "C"            "C"           
#>  [81,] "C"            "C"            "C"           
#>  [82,] "A"            "A"            "A"           
#>  [83,] "A"            "A"            "A"           
#>  [84,] "A"            "A"            "A"           
#>  [85,] "B"            "A"            "B"           
#>  [86,] "B"            "A"            "B"           
#>  [87,] "B"            "A"            "B"           
#>  [88,] "C"            "A"            "C"           
#>  [89,] "C"            "A"            "C"           
#>  [90,] "C"            "A"            "C"           
#>  [91,] "A"            "B"            "A"           
#>  [92,] "A"            "B"            "A"           
#>  [93,] "A"            "B"            "A"           
#>  [94,] "B"            "B"            "B"           
#>  [95,] "B"            "B"            "B"           
#>  [96,] "B"            "B"            "B"           
#>  [97,] "C"            "B"            "C"           
#>  [98,] "C"            "B"            "C"           
#>  [99,] "C"            "B"            "C"           
#> [100,] "A"            "C"            "A"           
#> [101,] "A"            "C"            "A"           
#> [102,] "A"            "C"            "A"           
#> [103,] "B"            "C"            "B"           
#> [104,] "B"            "C"            "B"           
#> [105,] "B"            "C"            "B"           
#> [106,] "C"            "C"            "C"           
#> [107,] "C"            "C"            "C"           
#> [108,] "C"            "C"            "C"           
#> [109,] "A"            "A"            "A"           
#> [110,] "B"            "B"            "A"           
#> [111,] "C"            "C"            "A"           
#> [112,] "A"            "A"            "A"           
#> [113,] "B"            "B"            "A"           
#> [114,] "C"            "C"            "A"           
#> [115,] "A"            "A"            "A"           
#> [116,] "B"            "B"            "A"           
#> [117,] "C"            "C"            "A"           
#> [118,] "A"            "A"            "B"           
#> [119,] "B"            "B"            "B"           
#> [120,] "C"            "C"            "B"           
#> [121,] "A"            "A"            "B"           
#> [122,] "B"            "B"            "B"           
#> [123,] "C"            "C"            "B"           
#> [124,] "A"            "A"            "B"           
#> [125,] "B"            "B"            "B"           
#> [126,] "C"            "C"            "B"           
#> [127,] "A"            "A"            "C"           
#> [128,] "B"            "B"            "C"           
#> [129,] "C"            "C"            "C"           
#> [130,] "A"            "A"            "C"           
#> [131,] "B"            "B"            "C"           
#> [132,] "C"            "C"            "C"           
#> [133,] "A"            "A"            "C"           
#> [134,] "B"            "B"            "C"           
#> [135,] "C"            "C"            "C"           
#> [136,] "A"            "A"            "A"           
#> [137,] "A"            "B"            "A"           
#> [138,] "A"            "C"            "A"           
#> [139,] "B"            "A"            "A"           
#> [140,] "B"            "B"            "A"           
#> [141,] "B"            "C"            "A"           
#> [142,] "C"            "A"            "A"           
#> [143,] "C"            "B"            "A"           
#> [144,] "C"            "C"            "A"           
#> [145,] "A"            "A"            "B"           
#> [146,] "A"            "B"            "B"           
#> [147,] "A"            "C"            "B"           
#> [148,] "B"            "A"            "B"           
#> [149,] "B"            "B"            "B"           
#> [150,] "B"            "C"            "B"           
#> [151,] "C"            "A"            "B"           
#> [152,] "C"            "B"            "B"           
#> [153,] "C"            "C"            "B"           
#> [154,] "A"            "A"            "C"           
#> [155,] "A"            "B"            "C"           
#> [156,] "A"            "C"            "C"           
#> [157,] "B"            "A"            "C"           
#> [158,] "B"            "B"            "C"           
#> [159,] "B"            "C"            "C"           
#> [160,] "C"            "A"            "C"           
#> [161,] "C"            "B"            "C"           
#> [162,] "C"            "C"            "C"           
#> [163,] "A"            "A"            "A"           
#> [164,] "B"            "A"            "A"           
#> [165,] "C"            "A"            "A"           
#> [166,] "A"            "A"            "A"           
#> [167,] "B"            "A"            "A"           
#> [168,] "C"            "A"            "A"           
#> [169,] "A"            "A"            "A"           
#> [170,] "B"            "A"            "A"           
#> [171,] "C"            "A"            "A"           
#> [172,] "A"            "B"            "B"           
#> [173,] "B"            "B"            "B"           
#> [174,] "C"            "B"            "B"           
#> [175,] "A"            "B"            "B"           
#> [176,] "B"            "B"            "B"           
#> [177,] "C"            "B"            "B"           
#> [178,] "A"            "B"            "B"           
#> [179,] "B"            "B"            "B"           
#> [180,] "C"            "B"            "B"           
#> [181,] "A"            "C"            "C"           
#> [182,] "B"            "C"            "C"           
#> [183,] "C"            "C"            "C"           
#> [184,] "A"            "C"            "C"           
#> [185,] "B"            "C"            "C"           
#> [186,] "C"            "C"            "C"           
#> [187,] "A"            "C"            "C"           
#> [188,] "B"            "C"            "C"           
#> [189,] "C"            "C"            "C"           
#> [190,] "A"            "A"            "A"           
#> [191,] "A"            "A"            "A"           
#> [192,] "A"            "A"            "A"           
#> [193,] "B"            "A"            "A"           
#> [194,] "B"            "A"            "A"           
#> [195,] "B"            "A"            "A"           
#> [196,] "C"            "A"            "A"           
#> [197,] "C"            "A"            "A"           
#> [198,] "C"            "A"            "A"           
#> [199,] "A"            "B"            "B"           
#> [200,] "A"            "B"            "B"           
#> [201,] "A"            "B"            "B"           
#> [202,] "B"            "B"            "B"           
#> [203,] "B"            "B"            "B"           
#> [204,] "B"            "B"            "B"           
#> [205,] "C"            "B"            "B"           
#> [206,] "C"            "B"            "B"           
#> [207,] "C"            "B"            "B"           
#> [208,] "A"            "C"            "C"           
#> [209,] "A"            "C"            "C"           
#> [210,] "A"            "C"            "C"           
#> [211,] "B"            "C"            "C"           
#> [212,] "B"            "C"            "C"           
#> [213,] "B"            "C"            "C"           
#> [214,] "C"            "C"            "C"           
#> [215,] "C"            "C"            "C"           
#> [216,] "C"            "C"            "C"