Convert IBD partition to a unique string for hashing

This is used for building a hash table for p(y at marker m|IBD).

Usage

hash.IP(IP, gs)

Arguments

IP

List containing vectors of genotype names with each vector corresponding to an IBD cell.

gs

Vector containing all genotype names.

Value

String where the integers in the IBD membership vector have been converted to ASCII characters.

Examples

gs <- paste0("g", 1:3)
IP1 <- list(c("g1", "g3"), c("g2"))
IP2 <- list(c("g2"), c("g3", "g1"))
hash1 <- hash.IP(IP1, gs)
hash2 <- hash.IP(IP2, gs)
hash1 == hash2 # TRUE, even though the order is different
#> [1] TRUE