Upper bounds on posterior probabilities

Upper bounds on posterior probabilities of 1st-recurrence recrudescence and reinfection (rows) for various MOI vectors (columns) that sum to at most eight. These upper bounds feature in vignette("understand-graph-prior") — clickable online (offline link) — and vignette("intra-episode-siblings") — clickable online (offline link). For a single recurrence (MOI vectors of length two), upper bounds are induced by the prior. For more than one recurrence (MOI vectors of length three or more), upper bounds assume all but the first recurrence has data; these bounds are only valid for monoclonal episodes; see XXX. Two upper bounds for each recurrence state are reported: one where summation over graphs includes graphs with cliques of three or more intra-episode siblings, another where summation over graphs excludes graphs with cliques of three or more intra-episode siblings. Since graphs with cliques of three or more intra-episode siblings only apply when MOI vectors include MOIs of three or more, these two bounds only differ for MOI vectors that include MOIs of three or more; see Examples.

Usage

maxima

Format

A matrix with 4 rows and 247 columns:

C_with

Row name; upper bound on the posterior probability of recrudescence when summation includes graphs with cliques of three or more intra-episode siblings.

C_wout

Row name; upper bound on the posterior probability of recrudescence when summation excludes graphs with cliques of three or more intra-episode siblings.

I_with

Row name; upper bound on the posterior probability of reinfection when summation includes graphs with cliques of three or more intra-episode siblings.

I_wout

Row name; upper bound on the posterior probability of reinfection when summation excludes graphs with cliques of three or more intra-episode siblings.

11

Column name; MOI vector (1, 1).

12

Column name; MOI vector (1, 2).

...
11111111

Column name; MOI vector (1, 1, 1, 1, 1, 1, 1, 1).

Source

https://github.com/aimeertaylor/Pv3Rs/blob/main/data-raw/maxima.R

Examples

# Convert column names to MOI character vectors
MOIvec <- strsplit(colnames(maxima), split = "")

# Get MOI vectors with MOIs that exceed two
Exceed2 <- sapply(MOIvec, function(x) any(as.numeric(x) > 2))

# Compare reinfection upper bounds with and with summation for MOI vectors that exclude
# and include MOIs greater than two
all(maxima["I_with", !Exceed2] == maxima["I_wout", !Exceed2])
#> [1] TRUE
any(maxima["I_with", Exceed2] == maxima["I_wout", Exceed2])
#> [1] FALSE