Generate parental allocations for a meiotic tetrad

For a given set of per chromosome marker counts, generate per marker parental ids for a meiotic tetrad generated by sexual recombination (chromosomal crossovers followed by independent assortment). We assume

For all chromosomes, both pairs of non-sister chromatids cross over

One crossover per non-sister chromatid pair

Equal-length chromosomes (follows from above; in reality, chromosomes have different lengths, and the number of crossovers increases with length)

Equi-distributed markers (implicit)

Usage

recombine_parent_ids(chrs_per_marker)

Arguments

chrs_per_marker

A vector of chromosome numbers for each marker.

Examples

n_chrs <- 14 # P. vivax has 14 chromosomes
n_markers <- 100 # For 100 markers
chrs_per_marker <- round(seq(0.51, n_chrs + 0.5, length.out = n_markers))
recombine_parent_ids(chrs_per_marker)
#>        [,1] [,2] [,3] [,4]
#>   [1,]    2    1    2    1
#>   [2,]    2    1    2    1
#>   [3,]    2    1    2    1
#>   [4,]    2    1    2    1
#>   [5,]    2    1    2    1
#>   [6,]    2    1    2    1
#>   [7,]    2    1    2    1
#>   [8,]    1    2    2    1
#>   [9,]    1    2    1    2
#>  [10,]    1    2    1    2
#>  [11,]    1    2    1    2
#>  [12,]    1    2    1    2
#>  [13,]    1    1    2    2
#>  [14,]    2    1    2    1
#>  [15,]    2    1    2    1
#>  [16,]    1    2    2    1
#>  [17,]    2    1    2    1
#>  [18,]    2    1    1    2
#>  [19,]    2    1    1    2
#>  [20,]    2    1    1    2
#>  [21,]    2    1    1    2
#>  [22,]    2    1    1    2
#>  [23,]    2    1    2    1
#>  [24,]    2    2    1    1
#>  [25,]    2    2    1    1
#>  [26,]    1    2    1    2
#>  [27,]    1    2    1    2
#>  [28,]    1    2    1    2
#>  [29,]    1    2    1    2
#>  [30,]    1    1    2    2
#>  [31,]    1    1    2    2
#>  [32,]    1    1    2    2
#>  [33,]    1    1    2    2
#>  [34,]    1    1    2    2
#>  [35,]    1    1    2    2
#>  [36,]    2    1    1    2
#>  [37,]    1    1    2    2
#>  [38,]    1    1    2    2
#>  [39,]    2    1    2    1
#>  [40,]    2    1    2    1
#>  [41,]    2    2    1    1
#>  [42,]    2    2    1    1
#>  [43,]    2    2    1    1
#>  [44,]    1    2    1    2
#>  [45,]    1    1    2    2
#>  [46,]    2    1    2    1
#>  [47,]    2    1    2    1
#>  [48,]    2    1    2    1
#>  [49,]    2    1    2    1
#>  [50,]    2    1    2    1
#>  [51,]    1    1    2    2
#>  [52,]    1    1    2    2
#>  [53,]    1    1    2    2
#>  [54,]    1    1    2    2
#>  [55,]    2    1    2    1
#>  [56,]    2    1    2    1
#>  [57,]    2    1    2    1
#>  [58,]    2    1    2    1
#>  [59,]    2    1    2    1
#>  [60,]    2    1    2    1
#>  [61,]    2    1    2    1
#>  [62,]    2    1    2    1
#>  [63,]    1    1    2    2
#>  [64,]    1    2    1    2
#>  [65,]    1    2    1    2
#>  [66,]    1    2    1    2
#>  [67,]    2    1    1    2
#>  [68,]    2    1    1    2
#>  [69,]    2    1    2    1
#>  [70,]    2    1    2    1
#>  [71,]    2    1    2    1
#>  [72,]    1    1    2    2
#>  [73,]    1    1    2    2
#>  [74,]    1    1    2    2
#>  [75,]    2    1    1    2
#>  [76,]    2    1    1    2
#>  [77,]    2    2    1    1
#>  [78,]    2    2    1    1
#>  [79,]    1    1    2    2
#>  [80,]    1    1    2    2
#>  [81,]    1    1    2    2
#>  [82,]    1    1    2    2
#>  [83,]    1    2    1    2
#>  [84,]    1    2    1    2
#>  [85,]    2    2    1    1
#>  [86,]    1    1    2    2
#>  [87,]    1    1    2    2
#>  [88,]    1    1    2    2
#>  [89,]    1    1    2    2
#>  [90,]    1    1    2    2
#>  [91,]    2    1    1    2
#>  [92,]    2    1    1    2
#>  [93,]    2    1    2    1
#>  [94,]    2    1    2    1
#>  [95,]    2    2    1    1
#>  [96,]    2    2    1    1
#>  [97,]    2    2    1    1
#>  [98,]    1    2    1    2
#>  [99,]    1    2    1    2
#> [100,]    1    2    1    2